refactor: rename repositories and update DTO usage for consistency
