--- src/main/java/com/gresk/modules/account/infrastructure/persistence/AccountJpaRepository.java
@@ -1,16 +1,0 @@
-package com.gresk.modules.account.infrastructure.persistence;
-
-import com.gresk.shared.domain.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface AccountJpaRepository extends JpaRepository<AccountEntity, UUID> {
-
-    Optional<AccountEntity> findByEmail(String email);
-
-    boolean existsByEmail(String email);
-
-    boolean existsByRolesContaining(Role role);
-}
